+++ api/src/controllers/announcement/deleteAnnouncement.js
@@ -1,0 +1,17 @@
+const { Announcement } = require('../../db');
+
+
+module.exports = async (req, res) => {
+    const { id } = req.params;
+  
+    try {
+      let act = await Announcement.destroy({
+        where: {
+          id: id,
+        },
+      });
+      return res.json({ erased: true });
+    } catch (err) {
+      return res.send({ error: err.message }).status(409);
+    }
+  }
