+++ api/src/controllers/review/getAnnouncementReviews.js
@@ -1,0 +1,19 @@
+const { Review } = require('../../db');
+
+module.exports = async (req, res) => {
+    const { id } = req.params;
+    try {
+      if (id) {
+        const review = await Review.findAll({
+          where: { announcementId: id },
+          include: {
+            model: User,
+            attributes: ["username"],
+          },
+        });
+        return res.json(review);
+      }
+    } catch (err) {
+      return res.send({ error: err.message }).status(409);
+    }
+  }
