+++ api/src/models/Favourite.js
@@ -1,7 +1,7 @@
 const { DataTypes } = require("sequelize");
 
 module.exports = (sequelize) => {
-  sequelize.define("favourites", {
+  sequelize.define("favourite", {
     id: {
       type: DataTypes.INTEGER,
       autoIncrement: true,
