+++ api/src/routes/index.js
@@ -4,20 +4,23 @@ const { Router } = require('express');
 // Ejemplo: const authRouter = require('./auth.js');
 
 
-const announcement = require("./announcement");
-const review = require("./review");
-const favourites = require("./favourites");
+// const announcement = require("./announcement");
+// const review = require("./review");
+// const favourites = require("./favourite");
 
 const router = Router();
 
 // Configurar los routers
 // Ejemplo: router.use('/auth', authRouter);
 const userRoutes = require('./user')
-router.use('/user', userRoutes);
+const announcementRoutes = require('./announcement')
+const reviewrRoutes = require('./review')
+const favouriteRoutes = require('./favourite')
 
-router.use("/announcement", announcement);
-router.use("/review", review);
-router.use("/favourites", favourites);
+router.use('/user', userRoutes);
+router.use("/announcement", announcementRoutes);
+router.use("/review", reviewrRoutes);
+router.use("/favourites", favouriteRoutes);
 
 module.exports = router
